legal-ner-api/requirements.txt
fastapi>=0.104.1
pydantic>=2.5.2
uvicorn[standard]>=0.24.0
torch>=2.6.0
transformers>=4.35.2
datasets>=2.14.0
psycopg2-binary>=2.9.9
redis>=5.0.1
minio>=7.2.0
python-multipart>=0.0.6
spacy>=3.7.2
aiofiles>=23.2.1
structlog>=23.2.0
prometheus-fastapi-instrumentator>=6.1.0
pydantic-settings>=2.0.0
sqlalchemy>=2.0.0
alembic>=1.13.0
tiktoken>=0.4.0
protobuf>=3.20.0
PyYAML>=6.0.1
sentence-transformers>=2.2.2
scikit-learn>=1.3.0
